separating cleaning from loading and saving
